Added basic regression test structure, improved Outcome/Variable parameters
Variables/Outcomes now strip all strings given as parameters. 

Regression test structure (documentation forthcoming) allows a set of regression tests to be run, in which they load a series of files of given data, expected results. The main code will need to be refactored more to allow for more testing, but it's worth it anyway.