Large CG Code Cleanup, menu generator added
Removed Queries from inference engine, renamed many variables/functions to be more intuitive, cleaned the menu setup to be shorter, added a menu generator option to take some list of items and make a menu and query the user to select one in it, reworked the do(X) computations to shorten that code by simply always doing a "list of sets" like when doing "all", and the config setting just specifies getting this list now. Also now, if a result is attempted to be cached and it is already cached BUT with a different value, an exception is raised.